+++ src/forsitan.hpp
@@ -23,4 +23,4 @@ extern Model* modelUlulo;
 extern Model* modelTabes;
 extern Model* modelLustro;
 extern Model* modelBulla;
-extern Model* modelContinua;
+extern Model* modelPerge;